Add features for OrderDetail

+++ backend/F-Sneaker/src/main/java/com/project/fsneaker/models/Order.java
@@ -57,7 +57,7 @@ public class Order {
     private String status;
 
     @Column(name = "total_money")
-    private Integer totalMoney;
+    private Float totalMoney;
 
     @Column(name = "shipping_method")
     private String shippingMethod;
